Tests/LibWeb/Text/expected/css/serialize-invalid-forgiving-selectors.txt

:is(hello, !?) -> :is(hello, !?)
:is(what, :fake(:imaginary), really) -> :is(what, :fake(:imaginary), really)
